+++ changelog.md
@@ -1,3 +1,34 @@
+# v0.0.50
+
+## Updates
+
+### Textfield
+
+- Textfield wrapper is now a Box.
+- Textfield props extend `BoxProps<HTMLElement>`.
+- Textfield props now explitictly contain `required`, `disabled`, and `inputProps`.
+- `inputProps` is for all properties, other than `required` and `disabled`, you'd like to pass to the `<input />` of Textfield.
+
+```ts
+export type TextFieldVariant = 'outlined' | 'filled';
+export type TextFieldInputProps = Partial<
+  Record<keyof svelte.JSX.HTMLAttributes<HTMLInputElement>, any>
+>;
+
+export interface TextFieldProps extends BoxProps<HTMLElement> {
+  variant: TextFieldVariant;
+  required?: boolean;
+  disabled?: boolean;
+  inputProps?: TextFieldInputProps;
+  label: string;
+  value?: string;
+  helperText?: string;
+  color?: ThemeColors;
+  error?: boolean;
+  characterCounter?: boolean;
+}
+```
+
 # v0.0.49
 
 ## Updates
@@ -8,7 +39,7 @@
 
 ### Menu
 
-- Changed `.menu` class  `position: absolute` to `position: fixed`
+- Changed `.menu` class `position: absolute` to `position: fixed`
 
 ### Checkbox